Update poll.ts
Idk if it will work (because I can't test it and I'm not familiar with TS) but this might stop the poll command from throwing an error every time it is run

+++ src/commands/utility/poll.ts
@@ -77,7 +77,9 @@ export const command: Command = {
                 }
             }
             if (!message.deleted) {
-                await message.delete().catch();
+                if (message.guild.me.hasPermission("MANAGE_MESSAGES")) {
+                    await message.delete().catch();
+                }
             }
         } catch (error) {
             xlg.error(error);
